fix(tests): fold the no-discarded-publish guard into the delivery-report suite
CI rejected the standalone file, and it was right to:

    [E] [PS-204 §2 orphan-test-file]
    tests/scitex_agent_container/a2a/test__no_discarded_publish.py:
    no matching src file (orphan test); mirror dir `src/scitex_agent_container/a2a`
    contains: _build.py, _card.py, ..., _delivery_report.py, ... — none match
    expected `_no_discarded_publish.py`

Every test file must mirror a source module. A cross-cutting invariant has no
module to mirror, so a file named after the invariant is unplaceable by
construction — the rule is not being obtuse, my file was.

The guard's natural home is beside the helper whose use it enforces:
`_delivery_report` exists so that a zero delivery is visible, and the guard
asserts that every publish actually routes through it. Same subject, one file.
The tests are unchanged — 15 pass, same as before the move — and the docstring
now records WHY they live here so the next person does not re-split them.

Local `tests/develop/test_audit.py` no longer reports PS-204 at all. It still
fails locally on findings that predate this branch (CLI-convention §4b/§6a/§13,
and a PS-224 on `.github/workflows/cla.yml` resolved against this container's
own `~/.scitex/dev/hosts.yaml`, which is not the registry CI uses). Those are
not mine to fold into this PR, and the quality-audit CI leg passed on the
previous push, so CI and this container disagree about that check — worth a
separate look, not a silent fix here.
